+++ R/anzg_boron_fresh.R
@@ -21,7 +21,8 @@
 #'\item{Genus}{The Genus name (chr).}
 #'\item{Group}{The taxonomic group (chr).}
 #'\item{Species}{The species binomial name (chr).}
-#'\item{Toxicity_measure}{The toxicity measure used (chr).} 
+#'\item{Toxicity_measure}{The toxicity measure used (chr).}
+#'\item{Units}{The concentration units of Conc (micrograms per Litre, ug/L) (chr).} 
 #'
 #' 
 #' Where toxicity measure is not a chronic NEC, EC10 or NOEC value,
@@ -33,7 +34,7 @@
 #' @name anzg_boron_fresh
 #' @docType data
 #' @format An object of class \code{tbl_df} (inherits from \code{tbl},
-#' \code{data.frame}) with 22 rows and 6 columns.
+#' \code{data.frame}) with 22 rows and 7 columns.
 #' @keywords datasets
 #' @examples
 #'