+++ R/anzg_copper_marine.R
@@ -24,7 +24,8 @@
 #'\item{Phylum}{The Phylum name (chr).}
 #'\item{Species}{The species binomial name (chr).}
 #'\item{Test_endpoint}{The test endpoint measure (chr).}
-#'\item{Toxicity_measure}{The toxicity measure used (chr).} 
+#'\item{Toxicity_measure}{The toxicity measure used (chr).}
+#'\item{Units}{The concentration units of Conc (micrograms per Litre, ug/L) (chr).} 
 #'
 #' 
 #' Where toxicity measure is not a chronic NEC, EC10 or NOEC value,
@@ -36,7 +37,7 @@
 #' @name anzg_copper_marine
 #' @docType data
 #' @format An object of class \code{tbl_df} (inherits from \code{tbl},
-#' \code{data.frame}) with 45 rows and 9 columns.
+#' \code{data.frame}) with 45 rows and 10 columns.
 #' @keywords datasets
 #' @examples
 #'