+++ R/anzg_data.R
@@ -64,7 +64,8 @@
 #'\item{Phylum}{The Phylum name (chr).}
 #'\item{Species}{The species binomial name (chr).}
 #'\item{Test_endpoint}{The test endpoint measure (chr).}
-#'\item{Toxicity_measure}{The toxicity measure used (chr).} 
+#'\item{Toxicity_measure}{The toxicity measure used (chr).}
+#'\item{Units}{The concentration units of Conc (micrograms per Litre, ug/L) (chr).} 
 #'
 #' 
 #' Where toxicity measure is not a chronic NEC, EC10 or NOEC value,
@@ -76,7 +77,7 @@
 #' @name anzg_data
 #' @docType data
 #' @format An object of class \code{tbl_df} (inherits from \code{tbl},
-#' \code{data.frame}) with 592 rows and 12 columns.
+#' \code{data.frame}) with 592 rows and 13 columns.
 #' @keywords datasets
 #' @examples
 #'