+++ R/csiro_cobalt_marine.R
@@ -11,13 +11,14 @@
 #' The columns are as follows:
 #' 
 #' \describe{ 
-#'\item{Conc}{The chemical concentration (dbl).}
+#'\item{Conc}{The chemical concentration in micrograms per Litre (dbl).}
 #'\item{Duration}{Test duration (chr).}
 #'\item{Group}{Taxonomic grouping information (chr).}
 #'\item{Life_stage}{Life stage of the test organism (chr).}
 #'\item{Species}{The species names name (chr).}
 #'\item{Test_endpoint}{Endpoint statistic, EC10, NEC etc (chr).}
-#'\item{Toxicity_measure}{Type of toxicity measure used (chr).} 
+#'\item{Toxicity_measure}{Type of toxicity measure used (chr).}
+#'\item{Units}{The concentration units of Conc (micrograms per Litre, ug/L) (chr).} 
 #'
 #' 
 #' Where toxicity measure is not a chronic NEC, EC10 or NOEC value,
@@ -31,7 +32,7 @@
 #' @name csiro_cobalt_marine
 #' @docType data
 #' @format An object of class \code{tbl_df} (inherits from \code{tbl},
-#' \code{data.frame}) with 14 rows and 7 columns.
+#' \code{data.frame}) with 14 rows and 8 columns.
 #' @keywords datasets
 #' @examples
 #'